Add: Add day 22 part 2 solution

+++ src/day22/mod.rs
@@ -11,7 +11,7 @@ use crate::models::AdventSolution;
 
 #[derive(Default)]
 pub struct Day22 {
-    parsed_data: Option<Vec<usize>>,
+    parsed_data: Option<Vec<i64>>,
 }
 
 impl AdventSolution for Day22 {
