+++ src/day22/parser.rs
@@ -1,3 +1,3 @@
-pub fn parse_input(input: String) -> Vec<usize> {
+pub fn parse_input(input: String) -> Vec<i64> {
     input.lines().map(|line| line.parse().unwrap()).collect()
 }
